--- rewrite-yaml/src/test/java/org/openrewrite/yaml/Issue5099Test.java
@@ -1,42 +1,0 @@
-/*
- * Copyright 2025 the original author or authors.
- * <p>
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- * <p>
- * https://www.apache.org/licenses/LICENSE-2.0
- * <p>
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.openrewrite.yaml;
-
-import org.intellij.lang.annotations.Language;
-import org.junit.jupiter.api.Test;
-import org.openrewrite.Issue;
-import org.openrewrite.test.RewriteTest;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.openrewrite.yaml.Assertions.yaml;
-
-class Issue5099Test implements RewriteTest {
-    @Issue("https://github.com/openrewrite/rewrite/issues/5099")
-    @Test
-    void parseFlowSequenceAtBufferBoundary() {
-        // May change over time in SnakeYaml, rendering this test fragile
-        var snakeYamlEffectiveStreamReaderBufferSize = 1024 - 1;
-
-        @Language("yml")
-        var yaml = "a: " + "x".repeat(1000) + "\n" + "b".repeat(16) + ": []";
-        assertEquals(snakeYamlEffectiveStreamReaderBufferSize - 1, yaml.lastIndexOf('['));
-
-        rewriteRun(
-            spec -> spec.recipe(new DeleteKey(".nonexistent","*")),
-            yaml(yaml)
-        );
-    }
-}
